js to cjs
ReferenceError: module is not defined in ES module scope
This file is being treated as an ES module because it has a '.js' file extension and 'C:\Users\aspan\Documents\GitHub\Udemy-React\04-gif-expert-app\package.json' contains "type": "module". To treat it as a CommonJS script, rename it to use the '.cjs' file extension.
    at file:///C:/Users/aspan/Documents/GitHub/Udemy-React/04-gif-expert-app/jest.config.js:1:1
    at ModuleJob.run (node:internal/modules/esm/module_job:194:25)